+++ examples/full_feature_set/file_management_implementation.h
