+++ examples/full_feature_set/main.c
@@ -27,7 +27,7 @@
 
 #include <openssl/ssl.h>
 
-#include "firmware_implementation.h"
+#include "file_management_implementation.h"
 #include "sensor_readings.h"
 
 #define DEFAULT_PUBLISH_PERIOD_SECONDS 15
